Use MUnit in ExecutionSuite
